second revisions

+++ backend/api/views.py
@@ -5,35 +5,33 @@ from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
 from recipes.models import (Favorite, Ingredient, IngredientAmount, Recipe,
                             ShoppingCart, Tag)
-from rest_framework import status, viewsets, permissions, response, decorators
+from rest_framework import status, viewsets, permissions, response
 from rest_framework.decorators import action
-from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 
-from .filters import RecipeFilter
-from .pagination import CustomPagination
-from .permissions import IsAdmin, IsAdminOrReadOnly
+from .filters import RecipeFilter, IngredientFilter
+from .pagination import LimitPagination
+from .permissions import IsAdminOrReadOnly, IsAuthenticated
 from .serializers import (
     FavoriteSerializer, IngredientSerializer,
     RecipeShortSerializer, RecipeWriteSerializer,
     ShoppingCartSerializer,
     SubscriptionUserSerializer, TagSerializer,
-    UserSingUpSerializer,
 )
 from users.models import Subscription, User
 
 
-class UserViewSet(UserViewSet):
+class CustomUserViewSet(UserViewSet):
     """Вьюсет для модели пользователя."""
 
     queryset = User.objects.all()
-    serializer_class = UserSingUpSerializer
-    pagination_class = CustomPagination
+    pagination_class = LimitPagination
+    permission_classes = [IsAdminOrReadOnly]
 
-    @decorators.action(
+    @action(
         detail=True,
         methods=['POST', 'DELETE'],
-        permission_classes=[permissions.IsAuthenticated]
+        permission_classes=[IsAuthenticated,]
     )
     def subscribe(self, request, **kwargs):
         user = request.user
@@ -49,28 +47,14 @@ class UserViewSet(UserViewSet):
             Subscription.objects.create(user=user, author=author)
             return response.Response(serializer.data,
                                      status=status.HTTP_201_CREATED)
-        get_object_or_404(Subscription, user=user, author=author).delete()
+        if request.method == 'DELETE':
+            get_object_or_404(Subscription, user=user, author=author).delete()
         return response.Response(status=status.HTTP_204_NO_CONTENT)
-
-    @decorators.action(
-        detail=False,
-        methods=['GET'],
-        permission_classes=(permissions.IsAuthenticated,)
-    )
-    @subscribe.mapping.delete
-    def unsubscribe(self, request, pk):
-        user = request.user
-        author = get_object_or_404(User, pk=pk)
-        Subscription.objects.filter(user=user, author=author).delete()
-        message = {
-            'detail': 'Вы успешно отписались'
-        }
-        return Response(message, status=status.HTTP_204_NO_CONTENT)
 
     @action(
         detail=False,
         methods=['GET'],
-        permission_classes=[IsAuthenticated]
+        permission_classes=[IsAuthenticated,]
     )
     def subscriptions(self, request):
         user = request.user
@@ -87,6 +71,7 @@ class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
     filter_backends = [DjangoFilterBackend]
+    filterset_class = IngredientFilter
 
 
 class TagViewSet(viewsets.ReadOnlyModelViewSet):
@@ -94,25 +79,24 @@ class TagViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
     permission_classes = (permissions.AllowAny,)
-    pagination_class = None
 
 
 class RecipeViewSet(viewsets.ModelViewSet):
     """Вьюсет для отображения рецептов."""
-    permission_classes = [IsAdminOrReadOnly, IsAdmin]
+    permission_classes = [IsAdminOrReadOnly, IsAuthenticated]
     filter_backends = [DjangoFilterBackend]
     filterset_class = RecipeFilter
 
+    @action(detail=True,)
     def get_serializer_class(self):
-        if self.action in ('favorite', 'shopping_cart'):
+        if self in ('favorite', 'shopping_cart'):
             return RecipeShortSerializer
-        if self.action in ('create', 'partial_update'):
+        elif self in ('create', 'partial_update'):
             return RecipeWriteSerializer
         return viewsets.ModelViewSet
 
     def get_queryset(self):
-        user_id = self.request.user.pk
-        return Recipe.objects.add_user_annotations(user_id).select_related(
+        return User.objects.select_related(
             'author'
         ).prefetch_related(
             'ingredients', 'tags'
@@ -121,7 +105,7 @@ class RecipeViewSet(viewsets.ModelViewSet):
     @action(
         detail=True,
         methods=['POST'],
-        permission_classes=[IsAdminOrReadOnly, IsAdmin]
+        permission_classes=[IsAdminOrReadOnly, IsAuthenticated]
     )
     def favorite(self, request, pk):
         user = request.user
@@ -149,7 +133,7 @@ class RecipeViewSet(viewsets.ModelViewSet):
     @action(
         detail=True,
         methods=['POST'],
-        permission_classes=[IsAdminOrReadOnly, IsAdmin]
+        permission_classes=[IsAuthenticated]
     )
     def shopping_cart(self, request, pk):
         user = request.user
@@ -180,22 +164,28 @@ class RecipeViewSet(viewsets.ModelViewSet):
         methods=['GET'],
         permission_classes=[IsAuthenticated]
     )
-    def download_shopping_cart(self):
-        return self.download_shopping_cart(self.request.user)
-
-    def shopping_cart(self, request):
+    def list_shopping_cart(self, request):
         ingredients = IngredientAmount.objects.filter(
             recipe__carts__user=request.user
         ).values(
             'ingredient__name', 'ingredient__measurement_unit'
-        ).annotate(ingredient_amount=Sum('amount'))
+        ).annotate(ingredient_amount=Sum('amount')
+                   ).values_list(
+            'ingredient__name', 'total_amount',
+            'ingredient__measurement_unit',
+        )
         shopping_list = ['Список покупок:\n']
         for ingredient in ingredients:
             name = ingredient['ingredient__name']
             unit = ingredient['ingredient__measurement_unit']
             amount = ingredient['ingredient_amount']
             shopping_list.append(f'\n{name} - {amount}, {unit}')
-        response = HttpResponse(shopping_list, content_type='text/plain')
+        return self.download_shopping_cart(self.request.user)
+
+    def download_shopping_cart(self):
+        response = HttpResponse(
+            self.list_shopping_cart, content_type='text/plain'
+        )
         response['Content-Disposition'] = \
             'attachment; filename="shopping_cart.txt"'
         return response
